Refactored how AuthContext works
User is now redirected automatically after creating an account. Facebook sign in added to AuthContext. auth.current user removed from Account.jsx because we should only get auth from one source.  We can get the current user from AuthContext. This is better because when the state in AuthContext changes it will cause a re render. If we do not trigger a re render when the user changes we will get a null user. When the user changes in the firebase.js it doesn't let the other components know something has changed. Keep your state in one place or it will be harder to keep track of your state.